cleanup one UT
